backend-python/database.py
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
import os
from dotenv import load_dotenv

load_dotenv()

# 默认使用 SQLite（无需安装数据库）
# 如需 MySQL，设置环境变量 DATABASE_URL=mysql+pymysql://user:password@host:port/database
DATABASE_URL = os.getenv("DATABASE_URL", None)

if DATABASE_URL:
    SQLALCHEMY_DATABASE_URL = DATABASE_URL
else:
    # 使用 SQLite，数据库文件保存在当前目录
    db_path = os.path.join(os.path.dirname(__file__), "ecosort.db")
    SQLALCHEMY_DATABASE_URL = f"sqlite:///{db_path}"

engine = create_engine(
    SQLALCHEMY_DATABASE_URL,
    connect_args={"check_same_thread": False} if "sqlite" in SQLALCHEMY_DATABASE_URL else {}
)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
